+++ src/pages/List.jsx
@@ -6,6 +6,13 @@ import { useTheme, styled } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 
+// Icons: Tags
+import OpenInNewIcon from '@mui/icons-material/OpenInNew';  // Link
+import DescriptionIcon from '@mui/icons-material/Description'; // Page
+import PodcastsIcon from '@mui/icons-material/Podcasts'; // Podcast
+import PlaceIcon from '@mui/icons-material/Place'; // Location
+import EventIcon from '@mui/icons-material/Event'; // Event
+
 // Custom UI
 import Loader from '../components/Loader';
 import StyledContent from '../components/StyledContent';
@@ -13,23 +20,25 @@ import StyledContent from '../components/StyledContent';
 // Helpers
 import {
   fetchDocument,
-  imageURL
 } from '../utils/utils';
 
-const style={
-  page: {
-    display: "flex",
-    flexDirection: "column",
-    height: "100%",
-    margin: "0",
-    padding: "0",
-  },
-  content: {
-    textAlign: "left",
-    padding: "0",
-    margin: "0",
-  },
-}
+const ListBox = styled(Box)(({ theme }) => ({
+  display: "flex",
+  flexDirection: "column",
+  gap: "2rem",
+  position: "relative",
+  marginTop: "0",
+  marginBottom: "0"
+}));
+
+const ListEntry = styled(Box)(({ theme }) => ({
+  display: "flex", position: "relative", gap: "1rem", width: "100%"
+}));
+
+const ListEntryTitle = styled(Typography)(({ theme }) => ({
+  margin: "0",
+  paddingBottom: "0",
+}));
 
 export default function List({listID}) {
   const theme = useTheme();
@@ -50,22 +59,50 @@ export default function List({listID}) {
     });
   }
 
+  const IconComponent = ({tagName}) => {
+    let icon = null;
+    switch(tagName) {
+      case "link":
+        icon = <OpenInNewIcon color="brand" style={{ pointerEvents: "none"}} />
+        break;
+      case "event":
+        icon = <EventIcon color="brand"  style={{ pointerEvents: "none"}} />
+        break;
+      case "page":
+        icon = <DescriptionIcon color="brand" style={{ pointerEvents: "none"}} />
+        break;
+      case "podcast":
+        icon = <PodcastsIcon color="brand" style={{ pointerEvents: "none"}}  />
+        break;
+      case "location":
+        icon = <PlaceIcon color="brand"  style={{ pointerEvents: "none"}} />
+        break;
+      default:
+        icon = <OpenInNewIcon color="brand"  style={{ pointerEvents: "none"}} />
+    }
+    return icon;
+  }
+
   return (
     <>
       { isLoading && <Loader />}
-      { !isLoading && <Box style={style.page} className="sff-list">
+      { !isLoading && <Box className="sff-list" style={{ marginTop: "2rem"}}>
         <StyledContent>
-          <Box style={style.content} className="scroll">
-            <Typography component="h1" variant='h_medium' style={{ marginTop: "0", marginBottom: "0"}}>
-              {listData.title}
-            </Typography>
+          <ListBox>
             { listData.items.map((item, index) => {
-              return <Box key={index}>
-                <Typography>{item.title}</Typography>
-                <ReactMarkdown>{item.content}</ReactMarkdown>
-              </Box>
+              return <ListEntry key={index}>
+                { item.url && <a href={item.url} target="_blank" rel="noreferrer"><IconComponent tagName={item?.tag} /></a> }
+                { !item.url && <IconComponent tagName={item?.tag} /> }
+                <Box style={{ display: "flex", flexDirection: "column", position: "relative", width: "100% "}}>
+                  <ListEntryTitle component="h2" style={{marginBottom: "0"}}>
+                    {item.title}
+                  </ListEntryTitle>
+                  <ReactMarkdown>{item.content}</ReactMarkdown>
+                  { item.url && <a href={item.url} target="_blank" rel="noreferrer">Link</a> }
+                </Box>
+              </ListEntry>
             })}
-          </Box>
+          </ListBox>
         </StyledContent>
       </Box>
       }